advances-v/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
grpcio==1.66.1
grpcio-tools==1.66.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
transformers
torch==2.4.1
bitsandbytes==0.43.3